Data.ProtocolBuffers.Internal

Documentation

type Tag = Word32

Field identifiers

data WireField

A representation of the wire format as described in https://developers.google.com/protocol-buffers/docs/encoding#structure

Constructors

VarintField !Tag !Word64
For: int32, int64, uint32, uint64, sint32, sint64, bool, enum

Fixed64Field !Tag !Word64
For: fixed64, sfixed64, double

DelimitedField !Tag !ByteString
For: string, bytes, embedded messages, packed repeated fields

StartField !Tag
For: groups (deprecated)

EndField !Tag
For: groups (deprecated)

Fixed32Field !Tag !Word32
For: fixed32, sfixed32, float

wireFieldTag :: WireField -> Tag

getWireField :: Get WireField

class EncodeWire a where

Methods

encodeWire :: Tag -> a -> Put

class DecodeWire a where

Methods

decodeWire :: WireField -> Get a

zzEncode32 :: Int32 -> Word32

zzEncode64 :: Int64 -> Word64

zzDecode32 :: Word32 -> Int32

zzDecode64 :: Word64 -> Int64

getVarintPrefixedBS :: Get ByteString

getVarInt :: (Integral a, Bits a) => Get a

putVarintPrefixedBS :: ByteString -> Put

putVarSInt :: (Integral a, Bits a) => a -> Put

This can be used on any Integral type and is needed for signed types; unsigned can use putVarUInt below. This has been changed to handle only up to 64 bit integral values (to match documentation).

putVarUInt :: (Integral a, Bits a) => a -> Put

This should be used on unsigned Integral types only (not checked)

newtype Field (n :: Nat) a

Fields are merely a way to hold a field tag along with its type, this shouldn't normally be referenced directly.

This provides better error messages than older versions which used Tagged

Constructors

Field

Fields

runField :: a

newtype Value a

Value selects the normal/typical way for encoding scalar (primitive) values.

Constructors

Value

Fields

runValue :: a

newtype Always a

To provide consistent instances for serialization a Traversable Functor is needed to make Required fields have the same shape as Optional, Repeated and Packed.

This is the Identity Functor with a Show instance.

Constructors

Always

Fields

runAlways :: a

newtype Enumeration a

Enumeration fields use fromEnum and toEnum when encoding and decoding messages.

Constructors

Enumeration

Fields

runEnumeration :: a

newtype RequiredField a

RequiredField is a newtype wrapped used to break overlapping instances for encoding and decoding values

Constructors

Required

Fields

runRequired :: a

newtype OptionalField a

OptionalField is a newtype wrapped used to break overlapping instances for encoding and decoding values

Constructors

Optional

Fields

runOptional :: a

newtype RepeatedField a

RepeatedField is a newtype wrapped used to break overlapping instances for encoding and decoding values

Constructors

Repeated

Fields

runRepeated :: a

newtype PackedField a

A Traversable Functor used to select packed sequence encoding/decoding.

Constructors

PackedField

Fields

runPackedField :: a

newtype PackedList a

A list that is stored in a packed format.

Constructors

PackedList

Fields

unPackedList :: [a]

newtype Message m

The way to embed a message within another message. These embedded messages are stored as length-delimited fields.

For example:

data Inner = Inner
   { innerField :: Required '1' (Value Int64)
   } deriving (Generic, Show)

 instance Encode Inner
instance Decode Inner

 data Outer = Outer
   { outerField :: Required '1' (Message Inner)
   } deriving (Generic, Show)

 instance Encode Outer
instance Decode Outer
 

It's worth noting that Message a is a Monoid and NFData instance. The Monoid behavior models that of the Protocol Buffers documentation, effectively Last. It's done with a fairly big hammer and it isn't possible to override this behavior. This can cause some less-obvious compile errors for paramterized Message types:

data Inner = Inner{inner :: Required '2' (Value Float)} deriving (Generic, Show)
instance Encode Inner
instance Decode Inner

data Outer a = Outer{outer :: Required '3' (Message a)} deriving (Generic, Show)
instance Encode a => Encode (Outer a)
instance Decode a => Decode (Outer a)
 

This fails because Decode needs to know that the message can be merged. The resulting error implies that you may want to add a constraint to the internal GMessageMonoid class:

/tmp/tst.hs:18:10:
  Could not deduce (protobuf-0.1:GMessageMonoid (Rep a))
    arising from a use of `protobuf-0.1: Decode .$gdmdecode'
  from the context (Decode a)
    bound by the instance declaration at /tmp/tst.hs:18:10-39
  Possible fix:
    add an instance declaration for
    (protobuf-0.1:GMessageMonoid (Rep a))
  In the expression:
    (protobuf-0.1:Decode.$gdmdecode)
  In an equation for decode:
      decode = (protobuf-0.1:Decode .$gdmdecode)
  In the instance declaration for `Decode (Outer a)'

The correct fix is to add the Monoid constraint for the message:

- instance (Encode a) => Decode (Outer a)
+ instance (Monoid (Message a), Decode a) => Decode (Outer a)

Constructors

Message

Fields

runMessage :: m

class GDecode (f :: * -> *)

Minimal complete definition

gdecode

class GEncode (f :: * -> *)

Minimal complete definition

gencode

class GMessageMonoid (f :: * -> *)

Minimal complete definition

gmempty, gmappend